Create Largest_Sum_Contiguous_Subarray.c
As a miner evaluating value stones while searching for ores, the miner wants to maximize their earnings by selecting a specific region to perform excavation operations. The mining area is divided into blocks and is assumed to be lined up in a one-dimensional plane. To find the contiguous blocks that will yield the miner's maximum profit and their amount of profit:

-> Design an algorithm in C language that finds the solution using a brute-force approach.
-> Design an algorithm in C language that finds the solution using a divide-and-conquer approach.