Finding the exit in the maze with Depth First Search
With the Depth First Search algorithm, an exit was found in the labyrinth given to us with a specified beginning and end.
Labyrinth read from file. While doing this, the x and y coordinates of the starting point were also kept in a pointer.
At the same time, a visitor matrix was created to mark and query the visited points.
In this matrix, a value of 1 for 'c' (for exit) was kept, a value of 0 for obstacles, a value of 2 for apples, and a value of 7 for wandered points and 3 for space.
In the Depth First Search function, if apple points were found, the score was increased by 10, if a dead end was reached, the score was decreased by 5.
As a start point 'b' in the maze was taken. Neighboring points that can be reached from this point were looked at. The process was continued in a recursive manner until all points were visited.