backtracking
On an NxN game board, N colors are placed in different order on each line. The order of the colors in a row can be changed by moving the colors circularly to the right. For example, if the colors in the row are red, blue, green, and purple, the new order will be purple, red, blue, green when the row is moved to the right 1 time. If it is shifted to the right one more time, green, purple, red, blue are obtained. In the last case, design the algorithm that recursively organizes the game board with the backtracking method so that there is only 1 of each color in each column.