Graf_Breath First Search
In this assignment, you are asked to develop an algorithm that detects people who are influencers in a social network. When designing your algorithm, use DFS and/or BFS algorithms where necessary.

There are different methods of detecting influencers on a social network. Two of these methods will be used in this assignment. In the directional graph given, the nodes show the names and surnames of the people, and the directional links show the links to the people that the person follows. This graph
The steps to be done on it are as follows:
1. When finding which nodes in a graph belong to influencer people, it is necessary to calculate for each node. But, especially in a large network, it will be unnecessary waste of time to trade for people with few followers. To eliminate nodes with no potential to become influencers, write the function that does the following:
Eliminate nodes whose in-degree value is less than a given M value by calculating the in-degree value of the nodes. After elimination, also eliminate nodes whose in-degree value falls below M. Repeat this process until there are no nodes with an in-degree value below M.
2. Write a function that calculates the total number of direct and indirect connections to a node for nodes not eliminated in step 1. For example, in the example below, apart from the direct connections to node 3 from nodes 4 and 5, there are a total of 4 connections from nodes 0, 4, 5 and 6 due to connections 0-5-3 and 0-6-4-3.